fixed deletefile

+++ src/gui/DeleteFileMenuItem.java
@@ -1,39 +1,46 @@
 package gui;
 
 
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 
-import javax.swing.JFileChooser;
+import javax.swing.JMenuItem;
 import javax.swing.JOptionPane;
 
 @SuppressWarnings("serial")
-public class DeleteFileMenuItem extends FileMenu {
-	private ToDo todo;
+public class DeleteFileMenuItem extends JMenuItem implements ActionListener{
+	private File file;
+	private ToDo todo; 
 	
 	public DeleteFileMenuItem(ToDo todo) {
-		super(todo, "Delete file");
+		super("Delete file");
 		this.todo = todo;
+		addActionListener(this);
 	}
 
+
 	@Override
-	protected void action(File file) throws FileNotFoundException {
+	public void actionPerformed(ActionEvent e) {
+		file = todo.getCurrentFile();
+		String message = "This will delete "+file.getName()+" entirely! Do you want to continue?";
+	   	 
+		 final String objButtons[] = { "Yes", "Cancel"};
+   	int option = JOptionPane.showOptionDialog(null,
+           	message, "Obs...",
+           	JOptionPane.DEFAULT_OPTION, JOptionPane.WARNING_MESSAGE, null,
+           	objButtons, objButtons[1]);
+   	if(option == 0){
 		try {
 			file.createNewFile();
 			file.delete();
-		} catch (IOException e) {
+		} catch (IOException e1) {
 			JOptionPane.showMessageDialog(null, "Something went wrong while deleting the file!", "Obs",
-					JOptionPane.INFORMATION_MESSAGE);
+					JOptionPane.ERROR_MESSAGE);
 			return;
 		}
-		
+   	}
+   	return;
 	}
-
-	@Override
-	protected int openDialog(JFileChooser fileChooser) {
-		fileChooser.setApproveButtonText("Delete");
-		return fileChooser.showOpenDialog(todo);
-	}
-
 }
